Move vendored dependency to `drools/vendor`
