+++ ansible_events/drools/README.md
@@ -1,8 +1,8 @@
 This module is based on original work by Jesus Ruiz
 found at https://github.com/jruizgit/rules
 
-This module has been adapted to substitute the
+The `vendor` module has been adapted to substitute the
 C library with a Python module that is redirecting
 the internal procedure invocation to a custom REST service.
 
-The REST service is implemented at https://github.com/mariofusco/drools-yaml-rules 
+The REST service is implemented at https://github.com/mariofusco/drools-yaml-rules
