--- ansible_events/drools/__init__.py
@@ -1,3 +1,0 @@
-import logging
-
-logger = logging.getLogger(__name__)
